Conflicts:
	source/stonehearth_data/mods/stonehearth/services/client/build_editor/build_editor_service.lua
	source/stonehearth_data/mods/stonehearth/ui/game/modes/build_mode/building_designer_2/building_designer_2.html
	source/stonehearth_data/mods/stonehearth/ui/game/modes/build_mode/building_designer_2/building_designer_2.js

+++ source/client/renderer/perfhud/render_context.cpp
@@ -31,10 +31,10 @@ RenderContext& RenderContext::SetTimelineHeight(perfmon::CounterValueType value)
 void RenderContext::DrawBox(csg::Rect2f const& r, csg::Color3 const& col)
 {
    float verts[] = {
-      r.min.x, r.min.y, 0, 1,
-      r.min.x, r.max.y, 0, 0,
-      r.max.x, r.max.y, 1, 0,
-      r.max.x, r.min.y, 1, 1
+      (float)r.min.x, (float)r.min.y, 0, 1,
+      (float)r.min.x, (float)r.max.y, 0, 0,
+      (float)r.max.x, (float)r.max.y, 1, 0,
+      (float)r.max.x, (float)r.min.y, 1, 1
    };
 
    h3dShowOverlays(verts, 4, col.r / 255.0f, col.g / 255.0f, col.b / 255.0f, 1.0f, material_.get(), 0);
@@ -43,18 +43,18 @@ void RenderContext::DrawBox(csg::Rect2f const& r, csg::Color3 const& col)
 void RenderContext::DrawString(std::string const& s, csg::Point2f const& pt, csg::Color3 const& col, csg::Color3 const& bg)
 {
    char const* text = s.c_str();
-   h3dutShowText(text, pt.x, pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
-   h3dutShowText(text, pt.x, pt.y + one_pixel_.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
-   h3dutShowText(text, pt.x + one_pixel_.x, pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
-   h3dutShowText(text, pt.x + one_pixel_.x + one_pixel_.y, pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
-   h3dutShowText(text, pt.x, pt.y, text_size_, col.r, col.g, col.b, font_material_.get());
+   h3dutShowText(text, (float)pt.x, (float)pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
+   h3dutShowText(text, (float)pt.x, (float)(pt.y + one_pixel_.y), text_size_, bg.r, bg.g, bg.b, font_material_.get());
+   h3dutShowText(text, (float)(pt.x + one_pixel_.x), (float)pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
+   h3dutShowText(text, (float)(pt.x + one_pixel_.x + one_pixel_.y), (float)pt.y, text_size_, bg.r, bg.g, bg.b, font_material_.get());
+   h3dutShowText(text, (float)pt.x, (float)pt.y, text_size_, col.r, col.g, col.b, font_material_.get());
 }
 
 void RenderContext::SetScreenSize(csg::Point2 const& size)
 {
    one_pixel_.x = 1.0f / size.x;
    one_pixel_.y = 1.0f / size.y;
-   text_size_ = one_pixel_.y * 20;
+   text_size_ = (float)(one_pixel_.y * 20);
 }
 
 csg::Point2f RenderContext::GetPixelSize()
