+++ source/csg/quaternion.h
@@ -13,8 +13,8 @@ class Quaternion
    public:
       // constructor/destructor
       inline Quaternion() : w(1.0f), x(0.0f), y(0.0f), z(0.0f) { }
-      inline Quaternion(float _w, float _x, float _y, float _z) : w(_w), x(_x), y(_y), z(_z) { }
-      Quaternion(Point3f const& axis, float angle);
+      inline Quaternion(double _w, double _x, double _y, double _z) : w(_w), x(_x), y(_y), z(_z) { }
+      Quaternion(Point3f const& axis, double angle);
       Quaternion(Point3f const& from, Point3f const& to);
       Quaternion(const protocol::quaternion& c) {
          LoadValue(c);
@@ -29,11 +29,11 @@ class Quaternion
       Quaternion& operator=(const Quaternion& other);
 
       // accessors
-      inline float& operator[](unsigned int i)         { return (&w)[i]; }
-      inline float operator[](unsigned int i) const    { return (&w)[i]; }
+      inline double& operator[](unsigned int i)         { return (&w)[i]; }
+      inline double operator[](unsigned int i) const    { return (&w)[i]; }
 
-      float magnitude() const;
-      float norm() const;
+      double magnitude() const;
+      double norm() const;
 
       void SaveValue(protocol::quaternion *msg) const;
       void LoadValue(const protocol::quaternion &msg);
@@ -46,16 +46,16 @@ class Quaternion
       bool is_identity() const;
 
       // manipulators
-      inline void set(float _w, float _x, float _y, float _z) {
+      inline void set(double _w, double _x, double _y, double _z) {
          w = _w; x = _x; y = _y; z = _z;
       }
-      void set(Point3f const& axis, float angle);
+      void set(Point3f const& axis, double angle);
       void set(Point3f const& from, Point3f const& to);
       // void set(const Matrix3& rotation);
-      void set(float z_rotation, float y_rotation, float x_rotation); 
+      void set(double z_rotation, double y_rotation, double x_rotation); 
       void lookAt(const Point3f& from, const Point3f& target);
 
-      void get_axis_angle(Point3f& axis, float& angle) const;
+      void get_axis_angle(Point3f& axis, double& angle) const;
 
       void clean();       // sets near-zero elements to 0
       void Normalize();   // sets to unit Quaternion
@@ -84,14 +84,14 @@ class Quaternion
       Quaternion operator-() const;
 
       // scalar multiplication
-      Quaternion&          operator*=(float scalar);
+      Quaternion&          operator*=(double scalar);
 
       // Quaternion multiplication
       Quaternion operator*(const Quaternion& other) const;
       Quaternion& operator*=(const Quaternion& other);
 
       // dot product
-      float Dot(const Quaternion& vector) const;
+      double Dot(const Quaternion& vector) const;
 
       // vector rotation
       Point3f rotate(Point3f const& vector) const;
@@ -102,24 +102,24 @@ class Quaternion
 
    public:
       // member variables
-      float w, x, y, z;      
+      double w, x, y, z;      
 
 };
 
 Quaternion conjugate(const Quaternion& quat);
 Quaternion inverse(const Quaternion& quat);
-float dot(const Quaternion& vector1, const Quaternion& vector2);
+double dot(const Quaternion& vector1, const Quaternion& vector2);
 
 // interpolation
-void lerp(Quaternion& result, const Quaternion& start, const Quaternion& end, float t);
-void slerp(Quaternion& result, const Quaternion& start, const Quaternion& end, float t);
-void approx_slerp(Quaternion& result, const Quaternion& start, const Quaternion& end, float t);
+void lerp(Quaternion& result, const Quaternion& start, const Quaternion& end, double t);
+void slerp(Quaternion& result, const Quaternion& start, const Quaternion& end, double t);
+void approx_slerp(Quaternion& result, const Quaternion& start, const Quaternion& end, double t);
 
-inline Quaternion operator*(float scalar, const Quaternion& q) {
+inline Quaternion operator*(double scalar, const Quaternion& q) {
    return Quaternion(scalar*q.w, scalar*q.x, scalar*q.y, scalar*q.z);
 }
 
-inline Quaternion Interpolate(const Quaternion &q0, const Quaternion &q1, float alpha) {
+inline Quaternion Interpolate(const Quaternion &q0, const Quaternion &q1, double alpha) {
    Quaternion result;
    slerp(result, q0, q1, alpha);
    return result;
