+++ source/csg/transform.cpp
@@ -7,9 +7,9 @@
 using namespace radiant;
 using namespace radiant::csg;
 
-csg::Transform csg::Transform::zero(Point3f(0.0f, 0.0f, 0.0f), Quaternion(0.0f, 0.0f, 0.0f, 1.0f));
+csg::Transform csg::Transform::zero(Point3f(0.0, 0.0, 0.0), Quaternion(0.0, 0.0, 0.0, 1.0));
 
-csg::Transform csg::Interpolate(const csg::Transform &t0, const csg::Transform &t1, float alpha)
+csg::Transform csg::Interpolate(const csg::Transform &t0, const csg::Transform &t1, double alpha)
 {
    csg::Transform t;
    t.position = csg::Interpolate(t0.position, t1.position, alpha);
