+++ source/physics/physics_util.cpp
@@ -32,11 +32,11 @@ Shape phys::LocalToWorld(Shape const& shape, om::EntityPtr entity)
    csg::Quaternion const& orientation = t.orientation;
 
    csg::Point3f axis;
-   float radAngle;
+   double radAngle;
    csg::GetAxisAngleNormalized(orientation, axis, radAngle);
 
    // Assumes rotation about the Y axis.
-   float degrees = radAngle * 180 / csg::k_pi;
+   double degrees = radAngle * 180 / csg::k_pi;
    int angle = (csg::ToClosestInt(degrees / 90) * 90) % 360;
 
    if (regionOrigin != csg::Point3f::zero) {
@@ -84,20 +84,20 @@ Shape phys::WorldToLocal(Shape const& shape, om::EntityPtr entity)
    return csg::ConvertTo<Shape::ScalarType, 3>(result);
 }
 
-bool phys::IsTerrainAligned(float modelOrigin)
+bool phys::IsTerrainAligned(double modelOrigin)
 {
-   float unused;
+   double unused;
    bool aligned = std::abs(std::modf(modelOrigin, &unused)) == 0.5f;
    return aligned;
 }
 
 // returns the translation needed to align the model boundaries to the terrain grid
-float phys::GetTerrainAlignmentOffset(float modelOrigin)
+double phys::GetTerrainAlignmentOffset(double modelOrigin)
 {
    if (phys::IsTerrainAligned(modelOrigin)) {
-      return 0.0f;
+      return 0.0;
    } else {
-      return -0.5f;
+      return -0.5;
    }
 }
